Refactors Learn content into a seperate file.
