+++ pubspec.yaml
@@ -50,10 +50,7 @@ dev_dependencies:
 # The following section is specific to Flutter.
 flutter:
   assets:
-    - assets/grass.JPG
-    - assets/skygrass.JPG
-    - assets/praireygrass.JPG
-    - assets/learn_background/customgrass.PNG
+    - assets/
     - assets/goal_progress/
     - assets/learn_background/
